Move InMemorySpanExporter to exporters/
