+++ sdk_contrib/testbed/build.gradle
@@ -4,6 +4,7 @@ dependencies {
     api project(':opentelemetry-api')
 
     implementation project(':opentelemetry-sdk'),
+            project(':opentelemetry-exporters-inmemory'),
             libraries.awaitility,
             libraries.guava
